- tables existance check added in migrations

+++ console/migrations/m130524_201442_init.php
@@ -12,22 +12,30 @@ class m130524_201442_init extends Migration
             $tableOptions = 'CHARACTER SET utf8 COLLATE utf8_unicode_ci ENGINE=InnoDB';
         }
 
-        $this->createTable('{{%user}}', [
-            'id' => $this->primaryKey(),
-            'username' => $this->string()->notNull()->unique(),
-            'auth_key' => $this->string(32)->notNull(),
-            'password_hash' => $this->string()->notNull(),
-            'password_reset_token' => $this->string()->unique(),
-            'email' => $this->string()->notNull()->unique(),
+        $tables = $this->db->schema->getTableNames();
 
-            'status' => $this->smallInteger()->notNull()->defaultValue(10),
-            'created_at' => $this->integer()->notNull(),
-            'updated_at' => $this->integer()->notNull(),
-        ], $tableOptions);
+        if (!in_array('user', $tables)) {
+
+            $this->createTable('{{%user}}', [
+                'id'            => $this->primaryKey(),
+                'username'      => $this->string()->notNull()->unique(),
+                'auth_key'      => $this->string(32)->notNull(),
+                'password_hash' => $this->string()->notNull(),
+                'password_reset_token' => $this->string()->unique(),
+                'email'         => $this->string()->notNull()->unique(),
+                'status'        => $this->smallInteger()->notNull()->defaultValue(10),
+                'created_at'    => $this->integer()->notNull(),
+                'updated_at'    => $this->integer()->notNull(),
+            ], $tableOptions);
+        }
     }
 
     public function down()
     {
-        $this->dropTable('{{%user}}');
+        $tables = $this->db->schema->getTableNames();
+
+        if (in_array('user', $tables)) {
+            $this->dropTable('{{%user}}');
+        }
     }
 }
