+++ console/migrations/m160422_100111_meta.php
@@ -7,23 +7,31 @@ class m160422_100111_meta extends Migration
 {
     public function safeUp()
     {
-        $tableOptions = 'ENGINE=InnoDB';
+        $tables = $this->db->schema->getTableNames();
 
-        $this->createTable(
-            '{{%meta}}',
-            [
-                'id'=> Schema::TYPE_PK."",
-                'name'=> Schema::TYPE_STRING."(50) NOT NULL COMMENT 'Tag name'",
+        if (!in_array('meta', $tables)) {
+            $tableOptions = 'ENGINE=InnoDB';
+
+            $this->createTable(
+                '{{%meta}}',
+                [
+                    'id'    => Schema::TYPE_PK . "",
+                    'name'  => Schema::TYPE_STRING . "(50) NOT NULL COMMENT 'Tag name'",
                 ],
-            $tableOptions
-        );
+                $tableOptions
+            );
 
-        $this->createIndex('meta_name_uindex', '{{%meta}}','name',1);
+            $this->createIndex('meta_name_uindex', '{{%meta}}', 'name', 1);
+        }
     }
 
     public function safeDown()
     {
-        $this->dropIndex('meta_name_uindex', '{{%meta}}');
-        $this->dropTable('{{%meta}}');
+        $tables = $this->db->schema->getTableNames();
+
+        if (in_array('meta', $tables)) {
+            $this->dropIndex('meta_name_uindex', '{{%meta}}');
+            $this->dropTable('{{%meta}}');
+        }
     }
 }
