+++ console/migrations/m160422_150311_question.php
@@ -7,40 +7,49 @@ class m160422_150311_question extends Migration
 {
     public function safeUp()
     {
-        $tableOptions = 'ENGINE=InnoDB';
+        $tables = $this->db->schema->getTableNames();
 
-        $this->createTable(
-            '{{%question}}',
-            [
-                'id'            => Schema::TYPE_PK." COMMENT 'ID'",
-                'title'         => Schema::TYPE_STRING."(500) NOT NULL COMMENT 'Question title'",
-                'descr'         => Schema::TYPE_TEXT." COMMENT 'Question text'",
-                'project_id'    => Schema::TYPE_INTEGER."(11) COMMENT 'Related project'",
-                'category_id'   => Schema::TYPE_INTEGER."(11) COMMENT 'Category'",
-                'language_id'   => Schema::TYPE_INTEGER."(11) COMMENT 'Programming language'",
-                'user_id'       => Schema::TYPE_INTEGER."(11) COMMENT 'Question author'",
-                'add_time'      => Schema::TYPE_DATETIME." COMMENT 'Add time'",
-                'edit_time'     => Schema::TYPE_DATETIME."",
-                'answer'        => Schema::TYPE_TEXT." COMMENT 'Authors answer'",
-                'error'         => Schema::TYPE_TEXT." COMMENT 'Error text'",
-            ],
-            $tableOptions
-        );
+        if (!in_array('question', $tables)) {
 
-        $this->createIndex('question_id_uindex', '{{%question}}','id',1);
-        $this->createIndex('fk_question',        '{{%question}}','project_id',0);
-        $this->createIndex('fk_category',        '{{%question}}','category_id',0);
-        $this->createIndex('fk_language',        '{{%question}}','language_id',0);
-        $this->createIndex('fk_user',            '{{%question}}','user_id',0);
+            $tableOptions = 'ENGINE=InnoDB';
+
+            $this->createTable(
+                '{{%question}}',
+                [
+                    'id'            => Schema::TYPE_PK . " COMMENT 'ID'",
+                    'title'         => Schema::TYPE_STRING . "(500) NOT NULL COMMENT 'Question title'",
+                    'descr'         => Schema::TYPE_TEXT . " COMMENT 'Question text'",
+                    'project_id'    => Schema::TYPE_INTEGER . "(11) COMMENT 'Related project'",
+                    'category_id'   => Schema::TYPE_INTEGER . "(11) COMMENT 'Category'",
+                    'language_id'   => Schema::TYPE_INTEGER . "(11) COMMENT 'Programming language'",
+                    'user_id'       => Schema::TYPE_INTEGER . "(11) COMMENT 'Question author'",
+                    'add_time'      => Schema::TYPE_DATETIME . " COMMENT 'Add time'",
+                    'edit_time'     => Schema::TYPE_DATETIME . "",
+                    'answer'        => Schema::TYPE_TEXT . " COMMENT 'Authors answer'",
+                    'error'         => Schema::TYPE_TEXT . " COMMENT 'Error text'",
+                ],
+                $tableOptions
+            );
+
+            $this->createIndex('question_id_uindex', '{{%question}}', 'id', 1);
+            $this->addForeignKey('fk_question_user_id', '{{%question}}', 'user_id', 'user', 'id');
+            $this->addForeignKey('fk_question_project_id', '{{%question}}', 'project_id', 'project', 'id');
+            $this->addForeignKey('fk_question_category_id', '{{%question}}', 'category_id', 'category', 'id');
+            $this->addForeignKey('fk_question_language_id', '{{%question}}', 'language_id', 'language', 'id');
+        }
     }
 
     public function safeDown()
     {
-        $this->dropIndex('question_id_uindex', '{{%question}}');
-        $this->dropIndex('fk_question', '{{%question}}');
-        $this->dropIndex('fk_category', '{{%question}}');
-        $this->dropIndex('fk_language', '{{%question}}');
-        $this->dropIndex('fk_user', '{{%question}}');
-        $this->dropTable('{{%question}}');
+        $tables = $this->db->schema->getTableNames();
+
+        if (in_array('question', $tables)) {
+            $this->dropIndex('question_id_uindex', '{{%question}}');
+            $this->dropForeignKey('fk_question_user_id', '{{%question}}');
+            $this->dropForeignKey('fk_question_project_id', '{{%question}}');
+            $this->dropForeignKey('fk_question_category_id', '{{%question}}');
+            $this->dropForeignKey('fk_question_language_id', '{{%question}}');
+            $this->dropTable('{{%question}}');
+        }
     }
 }
