+++ console/migrations/m160422_152811_project.php
@@ -7,27 +7,35 @@ class m160422_152811_project extends Migration
 {
     public function safeUp()
     {
-        $tableOptions = 'ENGINE=InnoDB';
+        $tables = $this->db->schema->getTableNames();
 
-        $this->createTable(
-            '{{%project}}',
-            [
-                'id'        => Schema::TYPE_PK . " COMMENT 'Project ID'",
-                'name'      => Schema::TYPE_STRING . "(255) NOT NULL COMMENT 'Project name'",
-                'descr'     => Schema::TYPE_TEXT . " COMMENT 'Description'",
-                'url'       => Schema::TYPE_STRING . "(200)",
-                'url_dev'   => Schema::TYPE_STRING . "(200)",
-                'url_prod'  => Schema::TYPE_STRING . "(200)",
-            ],
-            $tableOptions
-        );
+        if (!in_array('project', $tables)) {
+            $tableOptions = 'ENGINE=InnoDB';
 
-        $this->createIndex('project_name_uindex', '{{%project}}', 'name', 1);
+            $this->createTable(
+                '{{%project}}',
+                [
+                    'id'        => Schema::TYPE_PK . " COMMENT 'Project ID'",
+                    'name'      => Schema::TYPE_STRING . "(255) NOT NULL COMMENT 'Project name'",
+                    'descr'     => Schema::TYPE_TEXT . " COMMENT 'Description'",
+                    'url'       => Schema::TYPE_STRING . "(200)",
+                    'url_dev'   => Schema::TYPE_STRING . "(200)",
+                    'url_prod'  => Schema::TYPE_STRING . "(200)",
+                ],
+                $tableOptions
+            );
+
+            $this->createIndex('project_name_uindex', '{{%project}}', 'name', 1);
+        }
     }
 
     public function safeDown()
     {
-        $this->dropIndex('project_name_uindex', '{{%project}}');
-        $this->dropTable('{{%project}}');
+        $tables = $this->db->schema->getTableNames();
+
+        if (in_array('project', $tables)) {
+            $this->dropIndex('project_name_uindex', '{{%project}}');
+            $this->dropTable('{{%project}}');
+        }
     }
 }
