+++ console/migrations/m160428_153211_answer.php
@@ -7,29 +7,37 @@ class m160428_153211_answer extends Migration
 {
     public function safeUp()
     {
-        $tableOptions = 'ENGINE=InnoDB';
+        $tables = $this->db->schema->getTableNames();
 
-        $this->createTable(
-            '{{%answer}}',
-            [
-                'id'        => Schema::TYPE_PK." COMMENT 'Answer ID'",
-                'title'     => Schema::TYPE_STRING."(500) COMMENT 'Title'",
-                'descr'     => Schema::TYPE_TEXT." NOT NULL COMMENT 'Answer text'",
-                'user_id'   => Schema::TYPE_INTEGER."(11) COMMENT 'Author'",
-                'add_time'  => Schema::TYPE_DATETIME." COMMENT 'Answer time'",
-                'ref_url'   => Schema::TYPE_STRING."(500) COMMENT 'Reference URL'",
-            ],
-            $tableOptions
-        );
+        if (!in_array('answer', $tables)) {
+            $tableOptions = 'ENGINE=InnoDB';
 
-        $this->createIndex('answer_id_uindex', '{{%answer}}','id',1);
-        $this->createIndex('fk_question',      '{{%question}}', 'question_id', 0);
+            $this->createTable(
+                '{{%answer}}',
+                [
+                    'id' => Schema::TYPE_PK . " COMMENT 'Answer ID'",
+                    'title' => Schema::TYPE_STRING . "(500) COMMENT 'Title'",
+                    'descr' => Schema::TYPE_TEXT . " NOT NULL COMMENT 'Answer text'",
+                    'user_id' => Schema::TYPE_INTEGER . "(11) COMMENT 'Author'",
+                    'add_time' => Schema::TYPE_DATETIME . " COMMENT 'Answer time'",
+                    'ref_url' => Schema::TYPE_STRING . "(500) COMMENT 'Reference URL'",
+                ],
+                $tableOptions
+            );
+
+            $this->createIndex('answer_id_uindex', '{{%answer}}', 'id', 1);
+            $this->createIndex('fk_question', '{{%question}}', 'question_id', 0);
+        }
     }
 
     public function safeDown()
     {
-        $this->dropIndex('answer_id_uindex', '{{%question}}');
-        $this->dropIndex('fk_question', '{{%question}}');
-        $this->dropTable('{{%answer}}');
+        $tables = $this->db->schema->getTableNames();
+
+        if (in_array('answer', $tables)) {
+            $this->dropIndex('answer_id_uindex', '{{%question}}');
+            $this->dropIndex('fk_question', '{{%question}}');
+            $this->dropTable('{{%answer}}');
+        }
     }
 }
